Terrible giant blob commit... *sigh*
It turns out that enforcing good commit discipline when you're a 1-person
"team" is difficult and especially when you're writing the code to get
something done that you need access to immediately.

So, this is a terrible commit with all the changes that happened since last
time I worked on this project at least two weeks ago wherin I did who knows
what.  Who even knows if everything is in a working state right now?  (I think
it is, but hard to be sure *without any tests*!  Oy veh.)

Hopefully I learn my lesson this time...