Refactor bblearn.Filename class
The class was renamed to SubmissionFile to better reflect that it represents a
single file within a student's submission.  This will also fit better if
reading/copying facilities are added later, which they might be.

Additionally, all the properties were renamed for greater consistency.

Finally, the logic for actually retrieving the properties was moved into a
private classmethod which can be used to break up the BBLearn filename just
once at object creation time and a public classmethod was also provided so that
users of the class can decide whether they want to ask forgiveness or
permission when creating new objects.